requirements.txt
email-validator~=2.2.0
fastapi~=0.115.6
fastapi-mail~=1.4.2
Jinja2~=3.1.5
psycopg2-binary~=2.9.10
pydantic-settings~=2.7.1
python-jose~=3.3.0
python-multipart~=0.0.20
SQLAlchemy~=2.0.37
uvicorn~=0.34.0
Werkzeug~=3.1.3